pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cotlab"
version = "0.1.0"
description = "CoTLab - Chain of Thought Research Toolkit for LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Huseyin" }
]
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "vllm>=0.4.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "accelerate>=0.27.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.20.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[project.scripts]
cot-research = "cot_research.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cot_research"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
